move some scene config loading/processing to SceneLoader
