+++ src/scene.js
@@ -8,14 +8,13 @@ import Texture from './gl/texture';
 import VertexArrayObject from './gl/vao';
 import {StyleManager} from './styles/style_manager';
 import {StyleParser} from './styles/style_parser';
-import {mergeObjects} from './styles/rule';
+import SceneLoader from './scene_loader';
 import Camera from './camera';
 import Light from './light';
 import TileManager from './tile_manager';
 import DataSource from './sources/data_source';
 import FeatureSelection from './selection';
 import RenderState from './gl/render_state';
-import GLSL from './gl/glsl';
 
 import {Polygons} from './styles/polygons/polygons';
 import {Lines} from './styles/lines/lines';
@@ -942,141 +941,16 @@ export default class Scene {
             this.config_path = null;
         }
 
-        return Scene.loadScene(this.config_source, this.config_path).then(config => {
+        return SceneLoader.loadScene(this.config_source, this.config_path).then(config => {
             this.config = config;
-            this.preProcessConfig();
             this.trigger('load', { config: this.config });
             return this.config;
         });
     }
 
-    // Load scenes definitions from external URLs
-    // Optional *initial* path only (won't be passed to recursive 'include' calls)
-    // Useful for loading resources in base scene file from a separate location
-    // (e.g. in Tangram Play, when modified local scene should still refer to original resource URLs)
-    static loadScene(url, path = null) {
-        if (!url) {
-            return Promise.resolve({});
-        }
-
-        if (typeof url === 'string') {
-            path = path || Utils.pathForURL(url);
-        }
-
-        return Utils.loadResource(url).then(config => {
-            return StyleManager.loadRemoteStyles(config.styles, path).
-                then(styles => StyleManager.loadShaderBlocks(styles, path)). // TODO: deprecate remote shader blocks?
-                then(() => {
-                    // accept single-string or array
-                    if (typeof config.include === 'string') {
-                        config.include = [config.include];
-                    }
-
-                    if (!Array.isArray(config.include)) {
-                        Scene.normalizeTextures(config, path);
-                        return config;
-                    }
-
-                    // Collect URLs of scenes to include
-                    let includes = [];
-                    for (let url of config.include) {
-                        includes.push(Utils.addBaseURL(url, path));
-                    }
-                    delete config.include; // don't want to merge this property
-
-                    return Promise.
-                        all(includes.map(Scene.loadScene)).
-                        then(configs => {
-                            config = mergeObjects({}, ...configs, config);
-                            Scene.normalizeTextures(config, path);
-                            return config;
-                        });
-                });
-        });
-    }
-
-    // Expand paths and centralize texture definitions for a scene object
-    static normalizeTextures(config, path) {
-        config.textures = config.textures || {};
-
-        if (config.styles) {
-            for (let [style_name, style] of Utils.entries(config.styles)) {
-                if (style.texture) {
-                    // Texture by URL, expand relative to scene file
-                    if (typeof style.texture === 'string' && !config.textures[style.texture]) {
-                        style.texture = Utils.addBaseURL(style.texture, path);
-                    }
-                    // Texture by object, move it to the global scene texture set and give it a default name
-                    else if (typeof style.texture === 'object') {
-                        let texture_name = '__' + style_name;
-                        config.textures[texture_name] = style.texture;
-                        style.texture = texture_name; // point style to location of texture
-                    }
-                }
-
-                // If style has texture uniforms, expand texture URLs relative to scene file
-                if (style.shaders && style.shaders.uniforms) {
-                    for (let {type, value, key, uniforms} of GLSL.parseUniforms(style.shaders.uniforms)) {
-                        if (type === 'sampler2D' && !config.textures[value]) {
-                            uniforms[key] = Utils.addBaseURL(value, path);
-                        }
-                    }
-                }
-
-                // If style has material, expand texture URLs relative to scene file
-                if (style.material) {
-                    for (let prop of ['emission', 'ambient', 'diffuse', 'specular', 'normal']) {
-                        if (style.material[prop] != null &&
-                            style.material[prop].texture &&
-                            !config.textures[style.material[prop].texture]) {
-                            style.material[prop].texture = Utils.addBaseURL(style.material[prop].texture, path);
-                        }
-                    }
-                }
-            }
-        }
-
-        // Add path to textures
-        if (config.textures) {
-            for (let texture of Utils.values(config.textures)) {
-                texture.url = Utils.addBaseURL(texture.url, path);
-            }
-        }
-    }
-
-    // Normalize some settings that may not have been explicitly specified in the scene definition
-    preProcessConfig() {
-        // Assign ids to data sources
-        let source_id = 0;
-        for (let source in this.config.sources) {
-            this.config.sources[source].id = source_id++;
-        }
-
-        // If only one camera specified, set it as default
-        this.config.cameras = this.config.cameras || {};
-        if (this.config.camera) {
-            this.config.cameras.default = this.config.camera;
-        }
-        let camera_names = Object.keys(this.config.cameras);
-        if (camera_names.length === 0) {
-            this.config.cameras.default = { active: true };
-
-        }
-        else if (!this._active_camera) {
-            // If no camera set as active, use first one
-            this.config.cameras[camera_names[0]].active = true;
-        }
-
-        this.config.lights = this.config.lights || {}; // ensure lights object
-        this.config.styles = this.config.styles || {}; // ensure styles object
-
-        return this.config;
-    }
-
     loadDataSources() {
         for (var name in this.config.sources) {
             let source = this.config.sources[name];
-            source.url = Utils.addBaseURL(source.url, this.config_path);
             this.sources[name] = DataSource.create(Object.assign({}, source, {name}));
 
             if (!this.sources[name]) {
